practice_projects/react-agents-langgraph/pyproject.toml
[project]
name = "react-agents-langgraph"
version = "0.1.0"
description = ""
authors = [
    {name = "Rafael Gutierrez",email = "abadon.gutierrez@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0.0"
dependencies = [
    "langgraph (>=1.0.1,<2.0.0)",
    "langchain-openai (>=1.0.1,<2.0.0)",
    "langchain-tavily (>=0.2.12,<0.3.0)"
]

[tool.poetry]
packages = [{include = "react_agents_langgraph", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
